% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svine-array.R
\name{bimvine_mk}
\alias{bimvine_mk}
\title{Generate A Bivariate M-vine array}
\usage{
bimvine_mk(k)
}
\arguments{
\item{k}{Markov order k >= 1, default is k = 1, must be an integer}
}
\value{
A Bivariate M-vine array
}
\description{
Generate A Bivariate M-vine array with X_1, Y_1, ..., X_{k+1}, Y_{k+1}
on diagonal
}
\examples{
# return the R-vine array for Bivariate M-vine Markov order 2
bimvine_mk(2)

}
