% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvine-array.R
\name{extend_to_2d}
\alias{extend_to_2d}
\title{Helper function 2}
\usage{
extend_to_2d(mat)
}
\arguments{
\item{mat}{A 2dx2d vine matrix where the upper 1:d+1 columns have been filled.}
}
\value{
A filled 2dx2d vine matrix.
}
\description{
Helper function, extend column d+2:2d and corresponding rows
}
\examples{
d2_mk1 <- matrix(c(1, 1, 1, NA,
                  NA, 2, 2, NA,
                  NA, NA, 3, NA,
                  NA, NA, NA, 4), nrow = 4, ncol = 4, byrow = TRUE)
extend_to_2d(d2_mk1)

d3_mk1 = matrix(c(1, 1, 1, 1, NA,NA,
                 NA,2, 2, 2, NA,NA,
                 NA,NA,3, 3, NA,NA,
                 NA,NA,NA,4, NA,NA,
                 NA,NA,NA,NA,5, NA,
                 NA,NA,NA,NA,NA,6),
               nrow = 6, byrow = TRUE)
extend_to_2d(d3_mk1)

}
