% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvine-array.R
\name{fill_check}
\alias{fill_check}
\title{Helper function 3}
\usage{
fill_check(mat)
}
\arguments{
\item{mat}{A 2dx2d vine matrix where the upper 1:d+1 columns have been filled.}
}
\value{
A filled 2dx2d vine matrix if it is a valid svine array. NULL
             otherwise.
}
\description{
Helper function, extend column d+2:2d and corresponding rows
and then check whether the filled array is a valid svine array.
}
