% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_innov_func.R
\name{fit_bi_copula}
\alias{fit_bi_copula}
\title{Find the best fitted Bivariate copulas for consecutive points and/or bivariate data}
\usage{
fit_bi_copula(uts_1, uts_2 = NULL, print = FALSE, ...)
}
\arguments{
\item{uts_1}{univariate vector}

\item{uts_2}{univariate vector
Note: if null, them will separate uts_1 into 2 vectors}

\item{print}{whether to print the result, default to FALSE}

\item{...}{additional arguments pass to BiCopSelect}
}
\value{
all objects returned by BiCopSelect augmented with the following entries:
        C_{2|1}, C_{1|2}, and c12 of the fitted bivariate copula
}
\description{
Find the best fitted Bivariate copulas for consecutive points and/or bivariate data
}
\examples{
# set parameters
sim_par1 = matrix(c(0,0.5,0,0),2,2, byrow=T)
sim_fam = matrix(c(0,1,0,0),2,2, byrow=T)
rvine_ma = matrix(c(1,1,0,2),2,2, byrow=T)
RVM <- RVineMatrix(Matrix = rvine_ma, family = sim_fam, par = sim_par1)

# simulation
set.seed(123)
sim_sample = RVineSim(N = 5000, RVM = RVM)

# fit
fit_example = fit_bi_copula(uts_1 = sim_sample[,1], uts_2 = sim_sample[,2], print = TRUE)
fit_example$par
# Best fitting copula is Gaussian with tau 0.3304938 and AIC -1365.584
# 0.4961323

# note: the above result is same as following
cop1 <- BiCopSelect(sim_sample[,1], sim_sample[,2], familyset = 1:10, indeptest = FALSE, level = 0.05)
cop1$familyname
cop1$tau

}
