% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_innov_func.R
\name{fit_innov}
\alias{fit_innov}
\title{Fit d-variate Markov order k innovation model}
\usage{
fit_innov(dat, k = 1, print = FALSE, ...)
}
\arguments{
\item{dat}{matrix with number of columns = d; each column corresponds to a
univariate vector}

\item{k}{Markov order k >= 1, default is k = 1}

\item{print}{whether to print the result, default to FALSE}

\item{...}{additional arguments pass to BiCopSelect}
}
\value{
best fitted families, parameter estimates, total transitional loglikelihood
}
\description{
Find the best fitted Bivariate copulas for d-variate Markov order k
innovation model
}
\examples{
#source(file.path(proj, "simulation", "sim-bivariate-mk1.R"))

innov_dat = sim_bi_innov(n = 10000, family.innov = 1, par.innov = 0.7,
                         family.1 = 1, par.1 = 0.3,
                         family.2 = 1, par.2 = 0.6, trunc = 1000,
                         unif.margin = TRUE)

testfit = fit_innov(dat = innov_dat$dat, familyset = c(0,1,4,5,14))

test_uni = testfit$univariate_fit
test_innov = testfit$innov_fit

test_innov$family
test_innov$par

test_uni[[1]]$fitted_cop[[1]]$family
test_uni[[1]]$fitted_cop[[1]]$par
test_uni[[2]]$fitted_cop[[1]]$family
test_uni[[2]]$fitted_cop[[1]]$par

}
