% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_svine_func.R
\name{fit_svine}
\alias{fit_svine}
\title{Fit S-vine d-variate Markov order k model}
\usage{
fit_svine(dat, k = 1, ...)
}
\arguments{
\item{dat}{matrix with number of columns = d; each column corresponds to a
univariate vector}

\item{k}{Markov order k >= 1, default is k = 1}

\item{...}{additional arguments pass to RVineCopSelect}
}
\value{
best fitted families, parameter estimates, negative loglikelihood
}
\description{
Find the best fitted bivariate copulas for Markov order k, d-variate model
}
\examples{
# need to import sim_bi_mvine function from sim-bivariate-mk1.R
source(file.path(proj, "simulation", "sim-bivariate-mk1.R"))

# example for d=4, all Gaussian
apar1 = matrix(c(0, 0.4, 0.5, 0.4,
                 0, 0, 0.2, 0.3,
                 0, 0, 0, 0.2,
                 0, 0, 0, 0),
               4,4, byrow=T)
apar2 = matrix(0,4,4)
afam4 = matrix(1,4,4) # all Gaussian

# simulation
mvine_dat = sim_bi_mvine(5000,afam4,apar1,apar2)

testfit = fit_svine(dat = mvine_dat)
testfit[[1]]$Matrix
testfit[[1]]$family
testfit[[1]]$trans_logLik

}
