% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_svine_func.R
\name{fit_svine_single}
\alias{fit_svine_single}
\title{Fit a specific s-vine structure}
\usage{
fit_svine_single(Matrix, var_order, k = 1, utrain, utest, ...)
}
\arguments{
\item{var_order}{A permutation of numbers from 1 to d, re-arranges column in
utrain and utest}

\item{k}{Markov order k >= 1, default is k = 1}

\item{utrain}{matrix with number of columns = d; each column corresponds to a
univariate vector; training set}

\item{utest}{matrix with number of columns = d; each column corresponds to a
univariate vector; test set}

\item{...}{additional arguments pass to RVineCopSelect}
}
\value{
best fitted families, parameter estimates, negative loglikelihood
}
\description{
Return the fitted S-vine family and parameters, along with
            transitional loglikelihood on the training set and test set
}
\examples{

# example for d=4, all Gaussian
d2_mk1 = matrix(c(1,1,1,3,
                  0,2,2,1,
                  0,0,3,2,
                  0,0,0,4),
                  nrow = 4, byrow = TRUE)
apar1 = matrix(c(0, 0.4, 0.5, 0.4,
                 0, 0, 0.2, 0.3,
                 0, 0, 0, 0.2,
                 0, 0, 0, 0),
               4,4, byrow=T)
apar2 = matrix(0,4,4)
afam4 = matrix(1,4,4) # all Gaussian

# simulation
mvine_dat = sim_svine(n = 5000, vinematrix = d2_mk1, fam = afam4,
                      param1 = apar1, param2 = apar2)

testfit = fit_svine_single(Matrix = d2_mk1, var_order = c(1,2), k = 1,
                           utrain = mvine_dat, utest = mvine_dat,
                           familyset = c(0,1,4,5,14))
testfit$Matrix
testfit$family
testfit$trans_logLik
testfit$test_trans_logLik

}
