% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_innov_func.R
\name{fit_uni_mk}
\alias{fit_uni_mk}
\title{Find the best fitted Bivariate copulas for univariate Markov order k}
\usage{
fit_uni_mk(uts, k = 1, print = FALSE, separate = FALSE, ...)
}
\arguments{
\item{uts}{univariate vector}

\item{k}{Markov order k >= 1, default is k = 1}

\item{print}{whether to print the result, default to FALSE}

\item{separate}{Logical; whether log-likelihoods are returned point wisely
(default: separate = FALSE)}

\item{...}{additional arguments pass to BiCopSelect}
}
\value{
list of best fitted bivariate copulas, total transitional loglikelihood,
        and overall C_{2|1} (transition probability)
}
\description{
Find the best fitted Bivariate copulas for univariate Markov order k
}
\examples{
#source(file.path(proj, "simulation", "sim-univariate-mkp.R"))

## first example for d=4, all Gumbel
apar1 = matrix(c(0, 2,2,2, 0,0, 1.5,1.5, 0,0,0,1.8, 0,0,0,0),4,4, byrow=T)
apar2 = matrix(0,4,4)
afam4 = matrix(4,4,4) # all Gumbel

uts = sim_uni_mk(n=5000,fam=afam4,param1=apar1,param2=apar2)

# fit
fit_uts = fit_uni_mk(uts, k = 3, familyset = c(1,4,5,14), print = TRUE)

}
