% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvine-array.R
\name{generate_specific_upper_tri_matrices}
\alias{generate_specific_upper_tri_matrices}
\title{Helper function 1}
\usage{
generate_specific_upper_tri_matrices(mat, k)
}
\arguments{
\item{mat}{A 2dx2d vine matrix.}

\item{k}{kth Column, an integer.}
}
\value{
List of possible matrices.
}
\description{
Helper function, attempt to fill in all the cells in a given column
requires permn from Library combinat
}
\examples{
test <- matrix(c(1, 1, NA, NA,
                NA, 2, NA, NA,
                NA, NA, 3, NA,
                NA, NA, NA, 4), nrow = 4, ncol = 4, byrow = TRUE)
all_possible_matrices <- generate_specific_upper_tri_matrices(test, 3)
print(all_possible_matrices)

}
