% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_innov_func.R
\name{innov_out_sample_logLik}
\alias{innov_out_sample_logLik}
\title{Return the loglik on a given dataset based on the Markov order k innovation model}
\usage{
innov_out_sample_logLik(innov_model, umat)
}
\arguments{
\item{innov_model}{innovation in the form of the output of fit_innov function;
list of length 3 where first gives univariate fit, second element
gives innovations fit and last gives the transitional loglikelihood}

\item{umat}{matrix with number of columns = d; each column corresponds to a
univariate vector}
}
\value{
out-of-sample transitional loglikelihood on umat
}
\description{
Return the out of sample transitional loglikelhood of the test
             sample based on the Markov order k innovation model
}
\examples{
# example code
apar1x = matrix(c(0, 2,2,2, 0,0, 1.5,1.5, 0,0,0,1.2, 0,0,0,0),4,4, byrow=T)
apar2x = matrix(0,4,4)
afam4x = matrix(4,4,4) # all Gumbel

apar1y = matrix(c(0, 1.5,1.5,1.5, 0,0, 1.7,1.7, 0,0,0,2, 0,0,0,0),4,4, byrow=T)
apar2y = matrix(0,4,4)
afam4y = matrix(4,4,4) # all Gumbel

testing = sim_bi_innov_mkp(n = 5000, family.innov = 4, par.innov = 2,
                           family.1 = afam4x, par.1 = apar1x, par2.1 = apar2x,
                           family.2 = afam4y, par.2 = apar1y, par2.2 = apar2y,
                           unif.margin = TRUE)

testfit = fit_innov(dat = testing$dat, k=3, familyset = c(0,1,4,5,14))
testfit$trans_logLik
innov_out_sample_logLik(testfit, testing$dat)

}
