% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_svine_func.R
\name{is_lower_tri}
\alias{is_lower_tri}
\title{Check if a matrix is lower triangular}
\usage{
is_lower_tri(mat)
}
\arguments{
\item{mat}{triangular matrix}
}
\value{
Boolean, true if it is a lower triangular matrix
}
\description{
Check if a matrix is lower triangular; allow: upper triangular entries are
 of the same values (e.g. NA, 0, " ")
}
