% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_svine_func.R
\name{is_upper_tri}
\alias{is_upper_tri}
\title{Check if a matrix is upper triangular}
\usage{
is_upper_tri(mat)
}
\arguments{
\item{mat}{triangular matrix}
}
\value{
Boolean, true if it is a upper triangular matrix
}
\description{
Check if a matrix is upper triangular; allow: lower triangular entries are
 of the same values (e.g. NA, 0, " ")
}
