% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svine-array.R
\name{mk1_to_mkp}
\alias{mk1_to_mkp}
\title{Extend the given Markov order 1 M-vine array to Markov order p}
\usage{
mk1_to_mkp(mk1_mvine_array, p = 2)
}
\arguments{
\item{mk1_mvine_array}{matrix, Markov order 1 M-vine array, ncol = nrow and is even}

\item{p}{integer >= 1, Markov order, default 2}
}
\value{
A Markov order p M-vine array
}
\description{
Extend the given Markov order 1 M-vine array to Markov order p
}
\examples{
d2_mk1 = matrix(c(1,1,1,3,
                  NA,2,2,1,
                  NA,NA,3,2,
                  NA,NA,NA,4),
                nrow = 4, byrow = TRUE)
mk1_to_mkp(d2_mk1,p=2)  #Markov order 2
bimvine_mk(2)
mk1_to_mkp(d2_mk1,p=3)  #Markov order 3

d3_mk1 = matrix(c(1,1,1,1,4,4,
                  NA,2,2,2,1,5,
                  NA,NA,3,3,2,1,
                  NA,NA,NA,4,3,2,
                  NA,NA,NA,NA,5,3,
                  NA,NA,NA,NA,NA,6),
               nrow = 6, byrow = TRUE)
mk1_to_mkp(d3_mk1,p=2) #Markov order 2
mk1_to_mkp(d3_mk1,p=3) #Markov order 3

}
