% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svine-array.R
\name{num_to_xyz}
\alias{num_to_xyz}
\title{Replace numbers in a svine array to characters}
\usage{
num_to_xyz(mat, var_vec = c("X", "Y", "Z"), k = 1, upper.tri = TRUE)
}
\arguments{
\item{mat}{a d(k+1)xd(k+1) S-vine array}

\item{var_vec}{a vector of characters to replace; length d;
default c("X", "Y", "Z")}

\item{k}{Markov order k, default is k = 1, must be an integer}

\item{upper.tri}{Boolean, whether to convert the resulting matrix to upper
triangular; default TRUE}
}
\value{
a d(k+1)xd(k+1) S-vine array
}
\description{
Replace numbers in a d-dimensional Markov order k S-vine array
             with corresponding characters; assume 1 to d*(k+1) on diagonal
}
