% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_svine_func.R
\name{out_svine_par}
\alias{out_svine_par}
\title{Return averaged par matrix}
\usage{
out_svine_par(par, d = 2)
}
\arguments{
\item{par}{par matrix with inconsistent entries}

\item{d}{Markov order k >= 1, default is k = 1}
}
\value{
Averaged par matrix
}
\description{
Return averaged par matrix
}
