% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_svine_func.R
\name{return_consistent_model}
\alias{return_consistent_model}
\title{Return Consistent Model}
\usage{
return_consistent_model(Smodel, d = 2)
}
\arguments{
\item{Smodel}{list of length 4, must contain (Matrix, family, par, par2)}

\item{d}{number of variables, default is d = 2}
}
\value{
Consistent Model (list of length 4)
}
\description{
Return Consistent Model
}
