% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rvine-Rosenblatt-v2.R
\name{rotate_180}
\alias{rotate_180}
\title{Rotate a matrix by 180 degree}
\usage{
rotate_180(mat)
}
\arguments{
\item{mat}{a matrix to be rotated}
}
\value{
Returns the rotated matrix
}
\description{
Function to rotate a matrix by 180 degree
}
