% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svine-array.R
\name{svine_mk1}
\alias{svine_mk1}
\title{Generate all d-dimensional Markov order 1 S-vine array}
\usage{
svine_mk1(d = 2)
}
\arguments{
\item{d}{dimension k >= 2, default is k = 2, must be an integer}
}
\value{
List of 2dx2d S-vine arrays
}
\description{
Generate all possible d-dimensional Markov order 1 S-vine array
with 1=X_1, 2=X_2,...,d=X_d, d+1=X_1',...,2d=X_d' on diagonal
}
\examples{
svine_d2_list = svine_mk1(2)
svine_d3_list = svine_mk1(3)
svine_d4_list = svine_mk1(4)

}
