% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svine-array.R
\name{svine_mkp}
\alias{svine_mkp}
\title{Generate all d-dimensional Markov order p S-vine array}
\usage{
svine_mkp(d = 2, p = 1)
}
\arguments{
\item{d}{dimension k >= 2, default is k = 2, must be an integer}
}
\value{
List of d(p+1)xd(p+1) S-vine arrays
}
\description{
Generate all possible d-dimensional Markov order p S-vine array
             with 1=X_1, 2=X_2,...,d=X_d, d+1=X_1',...,2d=X_d' on diagonal
}
\examples{
svine_d3_p2_list = svine_mkp(d=3,p=2)

}
