% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_svine_func.R
\name{SVineCopSelect}
\alias{SVineCopSelect}
\title{Fit a S-vine model}
\usage{
SVineCopSelect(data, Matrix, k = 1, var_order = NULL, ...)
}
\arguments{
\item{data}{matrix with number of columns = d; each column corresponds to a
univariate vector with uniform margin}

\item{Matrix}{can be either:
1. lower or upper triangular 2d x 2d matrix that defines the
   Markov order 1 S-vine tree structure.
2. lower or upper triangular d(k+1) x d(k+1) matrix that defines the
   Markov order k S-vine tree structure.}

\item{k}{Markov order k >= 1, default is k = 1}

\item{var_order}{A permutation of numbers from 1 to d, re-arranges column in
data; default is 1:d (no re-arrangement needed)}

\item{...}{additional arguments pass to RVineCopSelect}
}
\value{
best fitted families, parameter estimates, loglikelihood
}
\description{
Return the fitted S-vine family and parameters, along with
            transitional loglikelihood on the dataset
}
