% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_svine_func.R
\name{SVineLogLik}
\alias{SVineLogLik}
\title{Calculates the log-likelihood of a S-vine copula model}
\usage{
SVineLogLik(data, Smodel)
}
\arguments{
\item{data}{matrix with number of columns = d; each column corresponds to a
univariate vector with uniform margin}

\item{Smodel}{list of at least length 5, must contain (Matrix, family, par, par2, var_order)}
}
\value{
log-likelihood
}
\description{
Calculates the log-likelihood of a S-vine copula model for a
             given copula data set.
}
\examples{
# example for d=4, all Gaussian
d2_mk1 = matrix(c(1,1,1,3,
                  0,2,2,1,
                  0,0,3,2,
                  0,0,0,4),
                nrow = 4, byrow = TRUE)
apar1 = matrix(c(0, 0.4, 0.5, 0.4,
                 0, 0, 0.2, 0.3,
                 0, 0, 0, 0.2,
                 0, 0, 0, 0),
               4,4, byrow=T)
apar2 = matrix(0,4,4)
afam4 = matrix(1,4,4) # all Gaussian

# simulation
set.seed(123)
mvine_dat = sim_svine(n = 5000, vinematrix = d2_mk1, fam = afam4,
#'                       param1 = apar1, param2 = apar2)

# fit
testfit = SVineCopSelect(data = mvine_dat, Matrix = d2_mk1, var_order = c(1,2),
                         k = 1, familyset = c(0,1,4,5,14))
summary(testfit)
contour(testfit)

testfit$Matrix
testfit$family
testfit$par
testfit$trans_logLik #0.2896287

}
