% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svine-array.R
\name{SVineMatrix_mk1_Check}
\alias{SVineMatrix_mk1_Check}
\title{Check whether the matrix is a valid Markov order 1 S-vine matrix}
\usage{
SVineMatrix_mk1_Check(M)
}
\arguments{
\item{M}{A 2dx2d vine matrix.}
}
\value{
code  1 for OK;
             -5 matrix is a valid RvineMatrix but not a valid Svine;
             -4 matrix is neither lower nor upper triangular;
             -3 diagonal can not be put in order d:1;
             -2 for not permutation of j:d in column d-j;
             -1 if cannot find proper binary array from array in natural order.
}
\description{
Testing whether the given upper triangular matrix is a valid
Markov order 1 S-vine matrix.
Requires function RVineMatrixCheck from package VineCopula.
}
\examples{
d2_mk1 = matrix(c(1,1,1,3,
                  NA,2,2,1,
                  NA,NA,3,2,
                  NA,NA,NA,4),
                nrow = 4, byrow = TRUE)
SVineMatrix_mk1_Check(d2_mk1) # should be 1

d2_mk1_F = matrix(c(1,1,1,2,
                    NA,2,2,1,
                    NA,NA,3,3,
                    NA,NA,NA,4),
                  nrow = 4, byrow = TRUE)
SVineMatrix_mk1_Check(d2_mk1_F) # should NOT be 1

d3_mk1 = matrix(c(1,1,1,1,4,4,
                  NA,2,2,2,1,5,
                  NA,NA,3,3,2,1,
                  NA,NA,NA,4,3,2,
                  NA,NA,NA,NA,5,3,
                  NA,NA,NA,NA,NA,6),
                nrow = 6, byrow = TRUE)
SVineMatrix_mk1_Check(d3_mk1)  # should be 1

d3_mk1_F = matrix(c(1,1,1,1,4,4,
                    NA,2,2,2,1,5,
                    NA,NA,3,3,3,1,
                    NA,NA,NA,4,2,2,
                    NA,NA,NA,NA,5,3,
                    NA,NA,NA,NA,NA,6),
                  nrow = 6, byrow = TRUE)
SVineMatrix_mk1_Check(d3_mk1_F)  # should NOT be 1

}
