% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_svine_func.R
\name{to_upper_tri}
\alias{to_upper_tri}
\title{Convert lower triangular matrix to upper triangular matrix}
\usage{
to_upper_tri(mat, NA_to_0 = TRUE)
}
\arguments{
\item{mat}{triangular matrix}

\item{NA_to_0}{Boolean, whether to convert NA to 0; default is TRUE}
}
\value{
Upper triangular matrix
}
\description{
Convert lower triangular matrix to upper triangular matrix and convert all
NA entries to 0 entries
}
