% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svine-array.R
\name{to_vinemat}
\alias{to_vinemat}
\title{Convert a svine array to svine conditional matrix}
\usage{
to_vinemat(mat)
}
\arguments{
\item{mat}{a d(k+1)xd(k+1) S-vine array}
}
\value{
a d(k+1)xd(k+1) S-vine array
}
\description{
Convert a svine array to svine conditional matrix; requires the
matrix to be upper triangular (will convert mat to upper triangular)
}
