% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_innov_func.R
\name{UnivariateLogLik}
\alias{UnivariateLogLik}
\title{Return the loglik on a given dataset based on the univariate Markov order k model}
\usage{
UnivariateLogLik(uni_model, uts, separate = FALSE, return_innov = FALSE)
}
\arguments{
\item{uni_model}{output returned by fit_uni_mk}

\item{uts}{univariate vector}

\item{separate}{Logical; whether log-likelihoods are returned point wisely
(default: separate = FALSE)}

\item{return_innov}{Logical; whether return innovations a = F_{k|1,...,k-1}
(default: return_innov = FALSE)}
}
\value{
out-of-sample transitional loglikelihood on the test set
}
\description{
Return the out of sample transitional loglikelhood of the test
             sample based on the univariate Markov order k model
}
\examples{
# Load necessary functions
source(file.path(proj, "simulation", "sim-univariate-mkp.R"))

## first example for d=2, all Gumbel
apar1 = matrix(c(0, 2, 0,0),2,2, byrow=T)
apar2 = matrix(0,2,2)
afam4 = matrix(4,2,2) # all Gumbel

uts_1 = sim_uni_mk(n=5000,fam=afam4,param1=apar1,param2=apar2)

# fit
fit_uts = fit_uni_mk(uts_1, k = 4, familyset = c(1,4,5,14), print = TRUE)

fit_uts$trans_logLik

# test function
out_llk = UnivariateLogLik(fit_uts, uts_1, separate = FALSE)
out_llk

## first example for d=4, all Gumbel
apar1 = matrix(c(0, 2,2,2, 0,0, 1.5,1.5, 0,0,0,1.8, 0,0,0,0),4,4, byrow=T)
apar2 = matrix(0,4,4)
afam4 = matrix(4,4,4) # all Gumbel

uts_3 = sim_uni_mk(n=5000,fam=afam4,param1=apar1,param2=apar2)

# fit
fit_uts = fit_uni_mk(uts_3, k = 3, familyset = c(1,4,5,14), print = TRUE)

fit_uts$trans_logLik

# test function
out_llk = UnivariateLogLik(fit_uts, uts_3, return_innov = TRUE)
out_llk

}
