% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rvine-Rosenblatt-v2.R
\name{varray2M}
\alias{varray2M}
\title{Convert a vine array to a maximum matrix}
\usage{
varray2M(A, iprint = FALSE, str = "")
}
\arguments{
\item{A}{a dxd vine array of R-vine in standard order with 1:d on diagonal, d>=2}

\item{iprint}{print flag for intermediate calculations}

\item{str}{string to describe vine for printing if iprint=T}
}
\value{
Returns a list with two components
\enumerate{
\item \code{M} - dxd array with m_{kj}= max a_{k1},..,a_{kj}
\item \code{icomp} - dxd indicator array on whether back step [k,j] is needed ; icomp[k-1,m_{kj}=1 if a_{kj}<m_{kj} for k>=2
}
}
\description{
Function to convert a vine array A to a maximum matrix. Algorithm 5 in Joe (2014).
}
