% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvine-array.R
\name{dvine}
\alias{dvine}
\title{Generate A D-vine array with 1:d on diagonal}
\usage{
dvine(d)
}
\arguments{
\item{d}{Number of variables, must be an integer}
}
\value{
A D-vine array with 1:d on diagonal
}
\description{
Generate A D-vine array with 1:d on diagonal
}
