% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_svine_func.R
\name{fit_svine}
\alias{fit_svine}
\title{Fit S-vine d-variate Markov order k model}
\usage{
fit_svine(dat, k = 1, train_prop = 1, Matrix = NULL, ...)
}
\arguments{
\item{dat}{matrix with number of columns = d; each column corresponds to a
univariate vector}

\item{k}{Markov order k >= 1, default is k = 1}

\item{train_prop}{Proportion of datasets in the training set, default is
train = 1 (i.e. no test set)}

\item{...}{additional arguments pass to RVineCopSelect}
}
\value{
best fitted families, parameter estimates, negative loglikelihood
}
\description{
Find the best fitted bivariate copulas for Markov order k, d-variate model
}
\examples{

# example for d=4, all Gaussian
d2_mk1 = matrix(c(1,1,1,3,
                  0,2,2,1,
                  0,0,3,2,
                  0,0,0,4),
                  nrow = 4, byrow = TRUE)
apar1 = matrix(c(0, 0.4, 0.5, 0.4,
                 0, 0, 0.2, 0.3,
                 0, 0, 0, 0.2,
                 0, 0, 0, 0),
               4,4, byrow=T)
apar2 = matrix(0,4,4)
afam4 = matrix(1,4,4) # all Gaussian

# simulation
mvine_dat = sim_svine(n = 5000, vinematrix = d2_mk1, fam = afam4,
                      param1 = apar1, param2 = apar2)

testfit = fit_svine(dat = mvine_dat, train_prop = 0.8, familyset = c(0,1,4,5,14))
testfit[[1]]$Matrix
testfit[[1]]$family
testfit[[1]]$trans_logLik

}
