% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rvine-Rosenblatt-v2.R, R/svine-array.R
\name{rotate_180}
\alias{rotate_180}
\title{Rotate a matrix by 180 degree}
\usage{
rotate_180(mat)

rotate_180(mat)
}
\arguments{
\item{mat}{Matrix}
}
\value{
Returns the rotated matrix

Same matrix except that it has been rotated by 180 degree
}
\description{
Function to rotate a matrix by 180 degree
}
